Unify DynamicTableTest
