+++ src/Shared/test/Shared.Tests/Http2/DynamicTableTest.cs
@@ -1,17 +1,100 @@
-﻿using System;
-using System.Collections;
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0.
+// See THIRD-PARTY-NOTICES.TXT in the project root for license information.
+
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Linq;
 using System.Net.Http.HPack;
 using System.Reflection;
 using System.Text;
-using System.Linq;
 using Xunit;
 
 namespace System.Net.Http.Unit.Tests.HPack
 {
     public class DynamicTableTest
     {
+        private readonly HeaderField _header1 = new HeaderField(Encoding.ASCII.GetBytes("header-1"), Encoding.ASCII.GetBytes("value1"));
+        private readonly HeaderField _header2 = new HeaderField(Encoding.ASCII.GetBytes("header-02"), Encoding.ASCII.GetBytes("value_2"));
+
+        [Fact]
+        public void DynamicTable_IsInitiallyEmpty()
+        {
+            DynamicTable dynamicTable = new DynamicTable(4096);
+            Assert.Equal(0, dynamicTable.Count);
+            Assert.Equal(0, dynamicTable.Size);
+            Assert.Equal(4096, dynamicTable.MaxSize);
+        }
+
+        [Fact]
+        public void DynamicTable_Count_IsNumberOfEntriesInDynamicTable()
+        {
+            DynamicTable dynamicTable = new DynamicTable(4096);
+
+            dynamicTable.Insert(_header1.Name, _header1.Value);
+            Assert.Equal(1, dynamicTable.Count);
+
+            dynamicTable.Insert(_header2.Name, _header2.Value);
+            Assert.Equal(2, dynamicTable.Count);
+        }
+
+        [Fact]
+        public void DynamicTable_Size_IsCurrentDynamicTableSize()
+        {
+            DynamicTable dynamicTable = new DynamicTable(4096);
+            Assert.Equal(0, dynamicTable.Size);
+
+            dynamicTable.Insert(_header1.Name, _header1.Value);
+            Assert.Equal(_header1.Length, dynamicTable.Size);
+
+            dynamicTable.Insert(_header2.Name, _header2.Value);
+            Assert.Equal(_header1.Length + _header2.Length, dynamicTable.Size);
+        }
+
+        [Fact]
+        public void DynamicTable_FirstEntry_IsMostRecentEntry()
+        {
+            DynamicTable dynamicTable = new DynamicTable(4096);
+            dynamicTable.Insert(_header1.Name, _header1.Value);
+            dynamicTable.Insert(_header2.Name, _header2.Value);
+
+            VerifyTableEntries(dynamicTable, _header2, _header1);
+        }
+
+        [Fact]
+        public void BoundsCheck_ThrowsIndexOutOfRangeException()
+        {
+            DynamicTable dynamicTable = new DynamicTable(4096);
+            Assert.Throws<IndexOutOfRangeException>(() => dynamicTable[0]);
+
+            dynamicTable.Insert(_header1.Name, _header1.Value);
+            Assert.Throws<IndexOutOfRangeException>(() => dynamicTable[1]);
+        }
+
+        [Fact]
+        public void DynamicTable_InsertEntryLargerThanMaxSize_NoOp()
+        {
+            DynamicTable dynamicTable = new DynamicTable(_header1.Length - 1);
+            dynamicTable.Insert(_header1.Name, _header1.Value);
+
+            Assert.Equal(0, dynamicTable.Count);
+            Assert.Equal(0, dynamicTable.Size);
+        }
+
+        [Fact]
+        public void DynamicTable_InsertEntryLargerThanRemainingSpace_NoOp()
+        {
+            DynamicTable dynamicTable = new DynamicTable(_header1.Length);
+            dynamicTable.Insert(_header1.Name, _header1.Value);
+
+            VerifyTableEntries(dynamicTable, _header1);
+
+            dynamicTable.Insert(_header2.Name, _header2.Value);
+
+            Assert.Equal(0, dynamicTable.Count);
+            Assert.Equal(0, dynamicTable.Size);
+        }
+
         [Theory]
         [InlineData(0)]
         [InlineData(1)]
@@ -20,8 +103,8 @@ namespace System.Net.Http.Unit.Tests.HPack
         public void DynamicTable_WrapsRingBuffer_Success(int targetInsertIndex)
         {
             FieldInfo insertIndexField = typeof(DynamicTable).GetField("_insertIndex", BindingFlags.NonPublic | BindingFlags.Instance);
-            var table = new DynamicTable(maxSize: 256);
-            var insertedHeaders = new Stack<byte[]>();
+            DynamicTable table = new DynamicTable(maxSize: 256);
+            Stack<byte[]> insertedHeaders = new Stack<byte[]>();
 
             // Insert into dynamic table until its insert index into its ring buffer loops back to 0.
             do
@@ -66,7 +149,7 @@ namespace System.Net.Http.Unit.Tests.HPack
             // This is purely to make it simple to perfectly reach our initial max size to test growing a full but non-wrapping buffer.
             Debug.Assert((insertSize % 64) == 0, $"{nameof(insertSize)} must be a multiple of 64 ({nameof(HeaderField)}.{nameof(HeaderField.RfcOverhead)} * 2)");
 
-            var dynamicTable = new DynamicTable(maxSize: initialMaxSize);
+            DynamicTable dynamicTable = new DynamicTable(maxSize: initialMaxSize);
             int insertedSize = 0;
 
             while (insertedSize != insertSize)
@@ -97,13 +180,76 @@ namespace System.Net.Http.Unit.Tests.HPack
             }
         }
 
+        [Fact]
+        public void DynamicTable_ResizingEvictsOldestEntries()
+        {
+            DynamicTable dynamicTable = new DynamicTable(4096);
+            dynamicTable.Insert(_header1.Name, _header1.Value);
+            dynamicTable.Insert(_header2.Name, _header2.Value);
+
+            VerifyTableEntries(dynamicTable, _header2, _header1);
+
+            dynamicTable.Resize(_header2.Length);
+
+            VerifyTableEntries(dynamicTable, _header2);
+        }
+
+        [Fact]
+        public void DynamicTable_ResizingToZeroEvictsAllEntries()
+        {
+            DynamicTable dynamicTable = new DynamicTable(4096);
+            dynamicTable.Insert(_header1.Name, _header1.Value);
+            dynamicTable.Insert(_header2.Name, _header2.Value);
+
+            dynamicTable.Resize(0);
+
+            Assert.Equal(0, dynamicTable.Count);
+            Assert.Equal(0, dynamicTable.Size);
+        }
+
+        [Fact]
+        public void DynamicTable_CanBeResizedToLargerMaxSize()
+        {
+            DynamicTable dynamicTable = new DynamicTable(_header1.Length);
+            dynamicTable.Insert(_header1.Name, _header1.Value);
+            dynamicTable.Insert(_header2.Name, _header2.Value);
+
+            // _header2 is larger than _header1, so an attempt at inserting it
+            // would first clear the table then return without actually inserting it,
+            // given it is larger than the current max size.
+            Assert.Equal(0, dynamicTable.Count);
+            Assert.Equal(0, dynamicTable.Size);
+
+            dynamicTable.Resize(dynamicTable.MaxSize + _header2.Length);
+            dynamicTable.Insert(_header2.Name, _header2.Value);
+
+            VerifyTableEntries(dynamicTable, _header2);
+        }
+
         public static IEnumerable<object[]> CreateResizeData()
         {
-            var values = new[] { 128, 256, 384, 512 };
+            int[] values = new[] { 128, 256, 384, 512 };
             return from initialMaxSize in values
                    from finalMaxSize in values
                    from insertSize in values
                    select new object[] { initialMaxSize, finalMaxSize, insertSize };
+        }
+
+        private void VerifyTableEntries(DynamicTable dynamicTable, params HeaderField[] entries)
+        {
+            Assert.Equal(entries.Length, dynamicTable.Count);
+            Assert.Equal(entries.Sum(e => e.Length), dynamicTable.Size);
+
+            for (int i = 0; i < entries.Length; i++)
+            {
+                HeaderField headerField = dynamicTable[i];
+
+                Assert.NotSame(entries[i].Name, headerField.Name);
+                Assert.Equal(entries[i].Name, headerField.Name);
+
+                Assert.NotSame(entries[i].Value, headerField.Value);
+                Assert.Equal(entries[i].Value, headerField.Value);
+            }
         }
     }
 }
